+++ src/id_translation/toml/factories/_cache_access.py
@@ -1,15 +1,16 @@
 from typing import Any
 
-from rics.misc import get_by_full_name
-
 from id_translation.fetching import CacheAccess
 from id_translation.types import IdType, SourceType
+
+from ._initialize import initialize
 
 
 def default_cache_access_factory(clazz: str, config: dict[str, Any]) -> CacheAccess[SourceType, IdType]:
     """Create a :class:`.CacheAccess` from config."""
-    cls = get_by_full_name(
+    return initialize(
         clazz,
-        subclass_of=CacheAccess,  # type: ignore[type-abstract]  # https://github.com/python/mypy/issues/4717
+        config,
+        CacheAccess,  # type: ignore[type-abstract]  # https://github.com/python/mypy/issues/4717
+        default_module=None,
     )
-    return cls(**config)